Fix encoding cross deps
