+++ pkg/nostale/encoding/nsl/nsl_test.go
@@ -1,8 +1,10 @@
-package auth
+package nsl
 
 import (
 	"bytes"
 	"testing"
+
+	"github.com/infinity-blackhole/elkia/pkg/nostale/encoding"
 )
 
 func TestDecoderDecode(t *testing.T) {
@@ -28,7 +30,7 @@ func TestDecoderDecode(t *testing.T) {
 			"BAE0870F40C2DA3666\n",
 	)
 	result := make([]byte, len(expected))
-	if err := NewDecoder(NewLoginEncoding(), &input).Decode(&result); err != nil {
+	if err := encoding.NewDecoder(NewEncoding(), &input).Decode(&result); err != nil {
 		t.Errorf("Error reading line: %s", err)
 	}
 	if !bytes.Equal(result, expected) {
@@ -44,7 +46,7 @@ func TestEncoderEncode(t *testing.T) {
 		130, 131, 10,
 	}
 	var buf bytes.Buffer
-	if err := NewEncoder(NewLoginEncoding(), &buf).Encode(input); err != nil {
+	if err := encoding.NewEncoder(NewEncoding(), &buf).Encode(input); err != nil {
 		t.Errorf("Unexpected error: %s", err)
 	}
 	result := buf.Bytes()
@@ -56,11 +58,11 @@ func TestEncoderEncode(t *testing.T) {
 func TestAsymmetricEncodeDecode(t *testing.T) {
 	expected := "fail Hello. This is a basic test"
 	var buff bytes.Buffer
-	if err := NewEncoder(NewLoginEncoding(), &buff).Encode(expected); err != nil {
+	if err := encoding.NewEncoder(NewEncoding(), &buff).Encode(expected); err != nil {
 		t.Errorf("Unexpected error: %s", err)
 	}
 	result := make([]byte, len(expected))
-	if err := NewDecoder(NewLoginEncoding(), &buff).Decode(&result); err == nil {
+	if err := encoding.NewDecoder(NewEncoding(), &buff).Decode(&result); err == nil {
 		t.Errorf("Expected error, got %s", result)
 	}
 }
