--- pkg/nostale/utils/utils.go
@@ -1,26 +1,0 @@
-package utils
-
-import (
-	"bufio"
-
-	eventing "github.com/infinity-blackhole/elkia/pkg/api/eventing/v1alpha1"
-	"github.com/infinity-blackhole/elkia/pkg/protonostale"
-	"github.com/sirupsen/logrus"
-)
-
-func WriteError(w *bufio.Writer, code eventing.DialogErrorCode, msg string) (n int, err error) {
-	if msg != "" {
-		logrus.Errorf("error: %s", msg)
-		if n, err = protonostale.WriteDialogInfoFrame(w, &eventing.DialogInfoFrame{
-			Content: msg,
-		}); err != nil {
-			return n, err
-		}
-	}
-	if n, err = protonostale.WriteDialogErrorFrame(w, &eventing.DialogErrorFrame{
-		Code: code,
-	}); err != nil {
-		return n, err
-	}
-	return n, nil
-}
